+++ web/src/app/dashboard/components/poll/PollForm.tsx
@@ -5,8 +5,8 @@ import {
   XMarkIcon,
 } from "@heroicons/react/24/outline";
 
+import { usePollForm } from "~/app/dashboard/hooks/usePollForm";
 import type { Poll } from "~/types";
-import { usePollForm } from "./usePollForm";
 
 type Props = {
   poll?: Poll;
